src/backend/requirements.txt
# Legal Document Analysis Backend Requirements
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI and ML dependencies
openai==1.3.7
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.2

# Document processing
pypdfium2==4.25.0
python-docx==1.1.0
pymupdf==1.23.8

# Data processing and utilities
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# HTTP and API utilities
httpx==0.25.2
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Background tasks and async
asyncio==3.4.3
aiofiles==23.2.1

# Logging and monitoring
loguru==0.7.2

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
